Even Faster yet
created list for primes, changed function when prime is found from print to append since interaction with list is faster then STDOUT. This is also more accurate since last revision skipped 1 and 2 as primes because i started loop with 3 and incremented by 2's. Now this list automatically includes 1 and 2 to start.